chore(pr) Added a test to cover the bug case.
